+++ mesos/mesos.go
@@ -79,7 +79,8 @@ func (e *Scheduler) EventLoop() {
 	res, err := e.Client.Do(e.Req)
 
 	if err != nil {
-		logrus.Fatal(err)
+		logrus.Error("Mesos Master connection error: ", err.Error())
+		return
 	}
 	defer res.Body.Close()
 
@@ -92,13 +93,18 @@ func (e *Scheduler) EventLoop() {
 
 	for {
 		// Read line from Mesos
-		line, _ = reader.ReadString('\n')
+		line, err = reader.ReadString('\n')
+		if err != nil {
+			logrus.Error("Error to read data from Mesos Master: ", err.Error())
+			return
+		}
 		line = strings.TrimSuffix(line, "\n")
 		// Read important data
 		var event mesosproto.Event // Event as ProtoBuf
 		err := jsonpb.UnmarshalString(line, &event)
 		if err != nil {
-			logrus.Error(err)
+			logrus.Error("Could not unmarshal Mesos Master data: ", err.Error())
+			continue
 		}
 		logrus.Debug("Subscribe Got: ", event.GetType())